(+) profile if asked

+++ examples/proto_bench/run.bash
@@ -3,6 +3,16 @@
 set -o noclobber
 GITROOT="../../"
 FROMGIT="examples/proto_bench"
+PROFILING=''
+
+while [[ "$1" =~ ^- && ! "$1" == "--" ]]; do
+    case $1 in
+        -p | --prof )
+            shift; PROFILING=yes
+            ;;
+    esac; shift
+done
+if [[ "$1" == '--' ]]; then shift; fi
 
 TIME_PATCH="timing.diff"
 PROF_PATCH="gprof.diff"
@@ -11,7 +21,9 @@ PROF_PATCH="gprof.diff"
 echo "Patching code"
 cd "$GITROOT"
 git apply "examples/proto_bench/$TIME_PATCH"
-git apply "examples/proto_bench/$PROF_PATCH"
+if [[ -n $PROFILING ]]; then
+    git apply "examples/proto_bench/$PROF_PATCH"
+fi
 
 echo "Compiling"
 make
@@ -21,16 +33,20 @@ lp="../../_build/install/default/bin/lambdapi"
 for t in *.lp; do
     echo "Checking $t..."
     $lp "$t"
-    echo "Profiling"
-    if [[ -e "$t"'.prof' ]]; then
-        rm "$t"'.prof'
+    if [[ -n $PROFILING ]]; then
+        echo "Profiling"
+        if [[ -e "$t"'.prof' ]]; then
+            rm "$t"'.prof'
+        fi
+        gprof "$lp" gmon.out > "$t"'.prof'
     fi
-    gprof "$lp" gmon.out > "$t"'.prof'
 done
 cd "$GITROOT"
 echo "Reverting patches"
 git apply --reverse "examples/proto_bench/$TIME_PATCH"
-git apply --reverse "examples/proto_bench/$PROF_PATCH"
+if [[ -n $PROFILING ]]; then
+    git apply --reverse "examples/proto_bench/$PROF_PATCH"
+fi
 
 ## Clean generatated files
 cd "$FROMGIT"
